+++ src/blockchain/executor.rs
@@ -820,7 +820,7 @@ impl<D: Db, W: Wm> Executor<D, W> {
         is_validator: bool,
         is_validator_closure: Arc<dyn IsValidator>,
     ) -> Result<Hash> {
-        debug!("EXEC BLOCK: {}", height);
+        debug!("Executing block: {}", height);
         // Write on a fork.
         let mut fork = self.db.write().fork_create();
 
@@ -945,13 +945,13 @@ impl<D: Db, W: Wm> Executor<D, W> {
                 ..
             }) => {
                 debug!("\tcan run height: {}", height);
-                debug!(
-                    "\t\t{:?}",
-                    hashes
-                        .iter()
-                        .all(|hash| matches!(pool.txs.get(hash), Some(Some(_))))
-                );
-                debug!("\t\t{:?}", hashes.iter());
+                //debug!(
+                //    "\t\t{:?}",
+                //    hashes
+                //        .iter()
+                //        .all(|hash| matches!(pool.txs.get(hash), Some(Some(_))))
+                //);
+                //debug!("\t\t{:?}", hashes.iter());
                 hashes
                     .iter()
                     .all(|hash| matches!(pool.txs.get(hash), Some(Some(_)))) // it might not put tcx in pool
@@ -970,7 +970,7 @@ impl<D: Db, W: Wm> Executor<D, W> {
             None => (Hash::default(), 0, 0),
         };
 
-        debug!("EXECUTOR:  last height: {}", height);
+        debug!("[executor] last height: {}", height);
 
         // TODO Maybe change seed here?
         #[allow(clippy::while_let_loop)]